refactor: 선택자 함수 util 폴더로 이동
