+++ src/js/renderMenuCount.js
@@ -1,4 +1,4 @@
-import { $ } from './selector.js'
+import { $ } from './util/selector.js'
 
 export const renderMenuCount = (menuCount) => {
   const $menuCount = $("#espresso-menu-count");
